langgrap-pr/langgraph-chatbot/backend/requirements.txt
langgraph>=0.2.23
langchain-core>=0.3.10
langchain-community>=0.3.5
langchain-groq>=0.1.3
python-dotenv>=1.0.1
fastapi>=0.110
uvicorn[standard]>=0.30
motor>=3.4
beautifulsoup4>=4.12
requests>=2.32
pydantic>=2.7
ddgs